objprint_dolphin objRenderShadow2 97.41->98.29: bit-reader address computed as (u8*)((pos >> 3) + (int)bs.data) - the int cast defeats pointer-base canonicalization, emitting srawi before the data load with target operand roles

+++ src/main/objprint_dolphin.c
@@ -2125,7 +2125,7 @@ void objRenderShadow2(int* obj, int* obj2, u8* m, int p4)
         {
             u32 w;
             int pos = bs.pos;
-            u8* p = (u8*)((pos >> 3) + bs.data);
+            u8* p = (u8*)((pos >> 3) + (int)bs.data);
             w = p[0];
             w |= p[1] << 8;
             w |= p[2] << 16;
@@ -2143,7 +2143,7 @@ void objRenderShadow2(int* obj, int* obj2, u8* m, int p4)
             {
                 u32 w;
                 int pos = bs.pos;
-                u8* p = (u8*)((pos >> 3) + bs.data);
+                u8* p = (u8*)((pos >> 3) + (int)bs.data);
                 w = p[0];
                 w |= p[1] << 8;
                 w |= p[2] << 16;
